Moved torch_device setting to root conftest
